+++ ui/lib/compliance/commons.ts
@@ -6,6 +6,7 @@ import { ENSCustomDetails } from "@/components/compliance/compliance-custom-deta
 import { GenericCustomDetails } from "@/components/compliance/compliance-custom-details/generic-details";
 import { ISOCustomDetails } from "@/components/compliance/compliance-custom-details/iso-details";
 import { KISACustomDetails } from "@/components/compliance/compliance-custom-details/kisa-details";
+import { MITRECustomDetails } from "@/components/compliance/compliance-custom-details/mitre-details";
 import { ThreatCustomDetails } from "@/components/compliance/compliance-custom-details/threat-details";
 import { AccordionItemProps } from "@/components/ui/accordion/Accordion";
 import {
@@ -42,6 +43,12 @@ import {
   toAccordionItems as toKISAAccordionItems,
 } from "./kisa";
 import {
+  calculateCategoryHeatmapData as calculateMITRECategoryHeatmapData,
+  getTopFailedSections as getMITRETopFailedSections,
+  mapComplianceData as mapMITREComplianceData,
+  toAccordionItems as toMITREAccordionItems,
+} from "./mitre";
+import {
   mapComplianceData as mapThetaComplianceData,
   toAccordionItems as toThetaAccordionItems,
 } from "./threat";
@@ -57,6 +64,7 @@ export interface ComplianceMapper {
     scanId: string | undefined,
   ) => AccordionItemProps[];
   getTopFailedSections: (mappedData: Framework[]) => FailedSection[];
+  calculateCategoryHeatmapData: (complianceData: Framework[]) => CategoryData[];
   getDetailsComponent: (requirement: Requirement) => React.ReactNode;
 }
 
@@ -103,6 +111,8 @@ const complianceMappers: Record<string, ComplianceMapper> = {
     mapComplianceData: mapENSComplianceData,
     toAccordionItems: toENSAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (data: Framework[]) =>
+      calculateCategoryHeatmapData(data),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(ENSCustomDetails, { requirement }),
   },
@@ -110,6 +120,8 @@ const complianceMappers: Record<string, ComplianceMapper> = {
     mapComplianceData: mapISOComplianceData,
     toAccordionItems: toISOAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (data: Framework[]) =>
+      calculateCategoryHeatmapData(data),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(ISOCustomDetails, { requirement }),
   },
@@ -117,6 +129,8 @@ const complianceMappers: Record<string, ComplianceMapper> = {
     mapComplianceData: mapCISComplianceData,
     toAccordionItems: toCISAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (data: Framework[]) =>
+      calculateCategoryHeatmapData(data),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(CISCustomDetails, { requirement }),
   },
@@ -124,6 +138,8 @@ const complianceMappers: Record<string, ComplianceMapper> = {
     mapComplianceData: mapAWSWellArchitectedComplianceData,
     toAccordionItems: toAWSWellArchitectedAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (data: Framework[]) =>
+      calculateCategoryHeatmapData(data),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(AWSWellArchitectedCustomDetails, { requirement }),
   },
@@ -131,6 +147,8 @@ const complianceMappers: Record<string, ComplianceMapper> = {
     mapComplianceData: mapAWSWellArchitectedComplianceData,
     toAccordionItems: toAWSWellArchitectedAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (data: Framework[]) =>
+      calculateCategoryHeatmapData(data),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(AWSWellArchitectedCustomDetails, { requirement }),
   },
@@ -138,13 +156,25 @@ const complianceMappers: Record<string, ComplianceMapper> = {
     mapComplianceData: mapKISAComplianceData,
     toAccordionItems: toKISAAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (data: Framework[]) =>
+      calculateCategoryHeatmapData(data),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(KISACustomDetails, { requirement }),
+  },
+  "MITRE-ATTACK": {
+    mapComplianceData: mapMITREComplianceData,
+    toAccordionItems: toMITREAccordionItems,
+    getTopFailedSections: getMITRETopFailedSections,
+    calculateCategoryHeatmapData: calculateMITRECategoryHeatmapData,
+    getDetailsComponent: (requirement: Requirement) =>
+      React.createElement(MITRECustomDetails, { requirement }),
   },
   ProwlerThreatScore: {
     mapComplianceData: mapThetaComplianceData,
     toAccordionItems: toThetaAccordionItems,
     getTopFailedSections,
+    calculateCategoryHeatmapData: (complianceData: Framework[]) =>
+      calculateCategoryHeatmapData(complianceData),
     getDetailsComponent: (requirement: Requirement) =>
       React.createElement(ThreatCustomDetails, { requirement }),
   },
@@ -155,6 +185,8 @@ const defaultMapper: ComplianceMapper = {
   mapComplianceData: mapGenericComplianceData,
   toAccordionItems: toGenericAccordionItems,
   getTopFailedSections,
+  calculateCategoryHeatmapData: (data: Framework[]) =>
+    calculateCategoryHeatmapData(data),
   getDetailsComponent: (requirement: Requirement) =>
     React.createElement(GenericCustomDetails, { requirement }),
 };